feat: Suricata監視の通知機能を強化（信頼度評価・Nmap強調・抑制サマリー対応）
- Suricataのアラート監視に信頼度評価 (confidence) を導入
- Severity=1 または Confidence=High の場合に @all 通知を自動付加
- Nmap偵察の初回のみ簡潔な緊急通知を送信、それ以降は suppress による抑制
- Signatureカテゴリに応じた絵文字を表示（SQL/SSH/NMAP等）
- suppressされたアラートは1分毎にサマリー通知として集約・報告
- Summary通知にも信頼度Lowを明示的に付加し、表記ブレを解消
