+++ mattermost_notifier/config/notice.py
@@ -2,7 +2,8 @@
 MATTERMOST_WEBHOOK_URL = "http://192.168.40.15:8065/hooks/gg5ijtkaf786tyccpu185ycf8o"
 
 # Suricataのeve.jsonログファイルパス
-EVE_JSON_PATH = "/var/log/suricata/eve.json"
+SURICATA_EVE_JSON_PATH = "/var/log/suricata/eve.json"
+
 # config/notice.py に追加
 OPENCANARY_LOG_PATH = "/opt/azazel/logs/opencanary.log"
